Initial commit
This is the initial commit of a Yeoman generator for PHP libraries that conform
to Wikimedia specifications, as affirmed by the RFC process and described in
<https://www.mediawiki.org/wiki/Manual:Developing_libraries>.

Known bug: if PSR-2 code style standard is selected, the generated project does
not pass its own lint check, because the autogenerated PHP files use the
MediaWiki style. I plan to solve this issue in a follow-up commit.

Bug: T104685
Change-Id: Ic8962b0f37dd68a7dc1bcce916bf946774cf74b0